third_party/blink/web_tests/fast/dom/shadow/select-in-shadowdom-expected.txt

When selecting from non-anchor Node to anchor node in ShadowDOM, client should not cause page jump.

Selecting from a node to another node in ShadowDOM. This should not start page navigation.
PASS lastClickTarget is host


Clicking a node in ShadowDOM.
PASS lastClickTarget is host


PASS successfullyParsed is true

TEST COMPLETE
